[*] move disc_fwd and gen_fwd from BaseTrainer to DCGANTrainer
